Filtering API rework

+++ Runtime/DataSet/Managed/CopyingDataSet.cs
@@ -54,7 +54,7 @@ namespace Massive
 				var blockOffset = blockIndex << 6;
 				while (block != 0UL)
 				{
-					var blockBit = deBruijn[(int)(((block & (ulong)-(long)block) * 0x37E84A99DAE458FUL) >> 58)];
+					var blockBit = (int)deBruijn[(int)(((block & (ulong)-(long)block) * 0x37E84A99DAE458FUL) >> 58)];
 
 					var bitsIndex = blockOffset + blockBit;
 					var bitsOffset = bitsIndex << 6;
@@ -68,7 +68,7 @@ namespace Massive
 					var bits = Bits[bitsIndex];
 					while (bits != 0UL)
 					{
-						var bit = deBruijn[(int)(((bits & (ulong)-(long)bits) * 0x37E84A99DAE458FUL) >> 58)];
+						var bit = (int)deBruijn[(int)(((bits & (ulong)-(long)bits) * 0x37E84A99DAE458FUL) >> 58)];
 						sourcePage[dataOffset + bit].CopyTo(ref destinationPage[dataOffset + bit]);
 						bits &= bits - 1UL;
 					}
